+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.8.8
+
+- Added `AtsMdfe` model
+- Added `mdfeId`, `mdfe` and `mdfes` values in `AtsOperation` model
+- Added default values to `StockClosingInput` model
+- Added `ATS_INVOICE_SWAP` value in `ReferenceCategory` enum
+
 ## 3.8.7
 
 - Updated `openingBalance` to `openingBalance20` in `StockClosingTank` model